collider cmp
Collider Cmp, it can be Circle or Rect. Collider has every necessary function. If the object is using CircleColl it will use that collision in Collides funct, otherwise, if the object is using RectColl it will use that collision in Collides func.
Right now the only collision possible is Circle