User Access Functions
Now the user can access to some TextRenderer functions to Initalize it, LoadFonts and DrawText. Now you don´t have to explicitly create the class in the main function.
I think it looks cleaner and easier since you just have to initialize the class like GLFW or GLAD. Also, the functions LoadFonts and DrawText is similar to the use of Text in the Raylib Library and I think is cleaner and easier